+++ pkg/meminfo/meminfo_unix_test.go
@@ -1,7 +1,7 @@
 //go:build linux || freebsd
 // +build linux freebsd
 
-package sysinfo
+package meminfo
 
 import (
 	"strings"
