+++ pkg/system/meminfo_deprecated.go
@@ -1,16 +1,16 @@
 package system
 
-import "github.com/docker/docker/pkg/sysinfo"
+import "github.com/docker/docker/pkg/meminfo"
 
 // MemInfo contains memory statistics of the host system.
 //
-// Deprecated: use [sysinfo.Memory].
-type MemInfo = sysinfo.Memory
+// Deprecated: use [meminfo.Memory].
+type MemInfo = meminfo.Memory
 
 // ReadMemInfo retrieves memory statistics of the host system and returns a
 // MemInfo type.
 //
-// Deprecated: use [sysinfo.ReadMemInfo].
-func ReadMemInfo() (*sysinfo.Memory, error) {
-	return sysinfo.ReadMemInfo()
+// Deprecated: use [meminfo.Read].
+func ReadMemInfo() (*meminfo.Memory, error) {
+	return meminfo.Read()
 }
